third_party/WebKit/LayoutTests/svg/custom/invalid-lengthlist-expected.txt

CONSOLE ERROR: line 13: Error: <text> attribute x: Expected length, "1.2.3".
